refactor: extract AdminIngestView's realtime subscription into a hook
