+++ frontend/components/admin/ingest.tsx
@@ -1,31 +1,9 @@
 "use client";
 
-import { useEffect, useState } from "react";
-import { isAxiosError } from "axios";
-
 import { NoticeCard } from "@components/ui/notice-card";
 import { Badge, type BadgeColor } from "@components/ui/badge";
 import { ButtonPrimary } from "@components/ui/button-primary";
-import { adminService } from "@services/admin";
-import { getSupabaseBrowserClient } from "@lib/supabase-browser";
-import { getApiErrorMessage } from "@utils/common";
-
-type JobStatus = "pending" | "running" | "done" | "failed";
-
-type JobResult = {
-  fetched?: number;
-  normalized?: number;
-  rejected?: number;
-} | null;
-
-type JobRow = {
-  id: string;
-  source: string;
-  status: JobStatus;
-  attempts: number;
-  result: JobResult;
-  error: string | null;
-};
+import { useIngestJob, type JobStatus } from "@hooks/use-ingest-job";
 
 const STATUS_COLOR: Record<JobStatus, BadgeColor> = {
   pending: "yellow",
@@ -34,62 +12,8 @@ const STATUS_COLOR: Record<JobStatus, BadgeColor> = {
   failed: "red",
 };
 
-const JOB_COLUMNS = "id, source, status, attempts, result, error";
-
-// Supabase query errors aren't AxiosErrors, so getApiErrorMessage falls back
-// to a generic message for them — surface their .message instead.
-const describeError = (e: unknown): string => {
-  if (isAxiosError(e)) return getApiErrorMessage(e);
-  if (e instanceof Error) return e.message;
-  return "Request failed.";
-};
-
 export const AdminIngestView = () => {
-  const [job, setJob] = useState<JobRow | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  const [triggering, setTriggering] = useState(false);
-
-  useEffect(() => {
-    if (!job) return;
-
-    const supabase = getSupabaseBrowserClient();
-    const channel = supabase
-      .channel(`jobs-${job.id}`)
-      .on(
-        "postgres_changes",
-        { event: "UPDATE", schema: "public", table: "jobs", filter: `id=eq.${job.id}` },
-        (payload) => setJob(payload.new as JobRow),
-      )
-      .subscribe();
-
-    return () => {
-      void supabase.removeChannel(channel);
-    };
-    // Re-subscribe only when the job id changes, not on every realtime update.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [job?.id]);
-
-  const triggerIngest = async () => {
-    setError(null);
-    setTriggering(true);
-    try {
-      const enqueued = await adminService.enqueueIngest();
-      const supabase = getSupabaseBrowserClient();
-      const { data, error: fetchError } = await supabase
-        .from("jobs")
-        .select(JOB_COLUMNS)
-        .eq("id", enqueued.job_id)
-        .single();
-      if (fetchError) throw fetchError;
-      setJob(data as JobRow);
-    } catch (e) {
-      setError(describeError(e));
-    } finally {
-      setTriggering(false);
-    }
-  };
-
-  const isBusy = job?.status === "pending" || job?.status === "running";
+  const { job, error, triggering, isBusy, triggerIngest } = useIngestJob();
 
   return (
     <div className="mx-auto max-w-screen-xl px-4 py-16 lg:py-20">
